Mahindra Lifespaces Nehru Nagar Pimpri Pune - Your Home, Your Choice
Pune is calling you to "Live a luxury life". The innovation in approach to constructing modern residential development is an excellent quality of Mahindra Pimpri Pune. From a magnificent structural design to passive interiors, it is a development to lead contented living.

Live safely & advantageously not sandwiched between bricks & mortar and loud road but by green surrounding and peaceful ambiance. Be advantaged, with everyday destination round the corner. With a first-class location and outstanding connectivity, life at this residential project in Pimpri, Pune.
Pimpri is a famous industrial and automotive hub of Pune. The growth in the industrial area drives residential development in the area mainly in the mid-income segment. It is surrounded by areas such as Vidya Nagar Pimple Saudagar and Tathawade. Some of the key zones in the locality include Bhosari, Vitthal Nagar, Nigdi, and Wakad. 

This residential Project is rich in what meets the heart. Luxurious Club House with swimming pool, gymnasium, indoor games, etc. Recreational area with jogging track, library, versatile hall, lounge, mini-theatre, and many other lifestyle amenities.

1 BHK, 2 BHK, and 3 BHK Residences in high-rise towers, Mahindra Lifespaces Nehru Nagar Pimpri Pune has a good number of green trees, flowering plants and open spaces to plunge you in tranquility and harmony. Involvement the unobstructed silence with modern facilities. Let the first-class typical living features treat you. These are all waiting for a faultless lifestyle.

The main purpose behind the launching of this development is to offer the heights of luxury living one can look for in Pune. In Nehru Nagar Pimpri, Mahindra Lifespaces presents all the latest avant-garde way of life services and amenities to go well with your preference for modern living. This residential project is a band of expensive lifestyle, which contains all your expectations from the dream home you have been looking for you and your kids.

For More Details :
Visit - https://mahindra.developerprojects.in/pune/mahindra-lifespaces-nehru-nagar-pimpri/